server/requirements.txt
sqlalchemy
fastapi
uvicorn
fastapi_health
alembic
asyncpg
pydantic-settings